fix: refactor recipe class into seperate file and removed it from recipe_ingredients_Page
